backend/requirements.txt
asgiref==3.8.1; python_version >= '3.8'
django==5.1.5; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
mysqlclient==2.1.1; python_version >= '3.8'  # For MySQL support
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
python-dotenv==0.21.1
pillow==10.3.0; python_version >= '3.8'  # For image handling
django-cors-headers==4.0.1; python_version >= '3.8' 
djangorestframework-simplejwt==5.2.2; python_version >= '3.8'  
